fix: UDP packet checksum computation
This fixes the following bugs:

RFC 768 states that the checksum is the 16-bit one's complement of
the one's complement sum of pseudo-header, UDP header, and data,
therefore the initial value of `0xffffu32` is erroneous, and should be
0 instead.

If payload data length is odd, the last call to
`buffer.read_u16::<BigEndian>()` will fail entirely, causing the last
byte of the payload to be ignored. Since that failure does not advance
the cursor's pointer, we can try again and read a u8 instead, then pad
it to 16-bit and use it in the checksum.